+++ src/contexts/ConnectRelayContext.js
@@ -1,15 +1,25 @@
-import React, { createContext, useState, useEffect, useContext } from 'react';
+import React, { createContext, useState, useEffect, useContext, useCallback } from 'react';
 import { ApiPromise, WsProvider } from "@polkadot/api";
 import { useLocalStorageContext } from './LocalStorageContext';
+import { useApiContextPara } from './ConnectParaContext'
 import useHealthCheck from '../hooks/useHealhCheck';
+import useApiSubscription from '../hooks/unSubHook';
+
+import { parseSchedule } from '../utilities/parseSchedule'
 
 const ApiContextRC = createContext();
 
 export function ApiConnectRC ({ children }) {
     const { network, restart } = useLocalStorageContext();
+    const { paraID } = useApiContextPara();
+
     const [api, setConnectedApi] = useState(null);
     const [isReady, setIsReady] = useState(false);
     const [provider, setProvider] = useState(null);
+    const [rcHeadInfo, setrcHeadInfo] = useState(null);
+    const [coretimeLeft, setCoretimeLeft] = useState(null);
+    const [paraHead, setParaHead] = useState(null);
+    const [paraCodeHash, setParaCodeHash] = useState(null);
 
     useEffect(() =>{
         const startApi = async () => {
@@ -19,7 +29,40 @@ export function ApiConnectRC ({ children }) {
         if(!provider && wsUri){
             startApi(wsUri);
         }
-    }, [network])
+        const getSchedule = async () => {
+          const schedule = await api.query.scheduler.agenda.entries();
+          const _coretimeLeft = parseSchedule(schedule, api, paraID)
+          setCoretimeLeft(_coretimeLeft)
+        }
+    
+        const getParaHead = async () => {
+          const _paraHead = await (await api.query.paras.heads(paraID)).toHuman()
+          setParaHead(_paraHead)
+        }
+    
+        const getParaCodeHash = async () => {
+          const _paraCodeHash = await (await api.query.paras.currentCodeHash(paraID)).toHuman()
+          setParaCodeHash(_paraCodeHash)
+        }
+    
+        if(api && paraID) {
+          getSchedule();
+          getParaHead();
+          getParaCodeHash();
+        }
+    }, [network, api, paraID, rcHeadInfo])
+
+    //Get relaychainHead information
+    const getNewRCHeads = useCallback(() => {
+        if(api){
+        return api.rpc.chain.subscribeNewHeads((lastHeader) => {
+            const head =  lastHeader.toHuman().number
+            setrcHeadInfo(head)
+        })
+        }
+    }, [api]);
+
+  useApiSubscription(getNewRCHeads, isReady);
 
     useHealthCheck(async ()=> {restart(); cleanupState()},network);
 
@@ -49,7 +92,7 @@ export function ApiConnectRC ({ children }) {
     }
 
     return (
-        <ApiContextRC.Provider value={{api, isReady}}>
+        <ApiContextRC.Provider value={{api, isReady, coretimeLeft, paraHead, paraCodeHash}}>
             { children }
         </ApiContextRC.Provider>
     );
