+++ src/utilities/handleHash.js
