+++ src/utilities/parseSchedule.js
