+++ src/views/dashboard/blockTableConfig.js
@@ -1,0 +1,12 @@
+export const blockColumns = [
+    {
+      key: 'head',
+      label: 'Number',
+      _props: { scope: 'col' },
+    },
+    {
+      key: 'headHash',
+      label:'Block Hash',
+      _props: { scope: 'col' },
+    },
+]
