+++ src/views/dashboard/collatorTableConfig.js
@@ -1,0 +1,17 @@
+export const collatorsColumns = [
+    {
+      key: 'name',
+      label: 'Name',
+      _props: { scope: 'col' },
+    },
+    {
+      key: 'address',
+      label:'Address',
+      _props: { scope: 'col' },
+    },
+    {
+      key: 'ws',
+      label:'WebSocket',
+      _props: { scope: 'col' },
+    },
+]
